Move repo/entity/dao to user package
