+++ stream-chat-android-offline/api/stream-chat-android-offline.api
@@ -243,14 +243,6 @@ public final class io/getstream/chat/android/livedata/dao/SyncStateDao_Impl : io
 	public fun select (Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
-public final class io/getstream/chat/android/livedata/dao/UserDao_Impl : io/getstream/chat/android/livedata/dao/UserDao {
-	public fun <init> (Landroidx/room/RoomDatabase;)V
-	public fun insert (Lio/getstream/chat/android/livedata/entity/UserEntity;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun insertMany (Ljava/util/List;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun select (Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-	public fun select (Ljava/util/List;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
-}
-
 public final class io/getstream/chat/android/livedata/extensions/ClientExtensionsKt {
 	public static final fun isPermanent (Lio/getstream/chat/android/client/errors/ChatError;)Z
 }
@@ -265,7 +257,7 @@ public final class io/getstream/chat/android/livedata/repository/database/ChatDa
 	public fun queryChannelsQDao ()Lio/getstream/chat/android/livedata/dao/QueryChannelsDao;
 	public fun reactionDao ()Lio/getstream/chat/android/livedata/repository/domain/reaction/ReactionDao;
 	public fun syncStateDao ()Lio/getstream/chat/android/livedata/dao/SyncStateDao;
-	public fun userDao ()Lio/getstream/chat/android/livedata/dao/UserDao;
+	public fun userDao ()Lio/getstream/chat/android/livedata/repository/domain/user/UserDao;
 }
 
 public final class io/getstream/chat/android/livedata/repository/domain/message/MessageDao_Impl {
@@ -296,6 +288,14 @@ public final class io/getstream/chat/android/livedata/repository/domain/reaction
 	public fun selectSyncNeeded (Lio/getstream/chat/android/client/utils/SyncStatus;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public fun selectUserReactionsToMessage (Ljava/lang/String;Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 	public fun setDeleteAt (Ljava/lang/String;Ljava/lang/String;Ljava/util/Date;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+}
+
+public final class io/getstream/chat/android/livedata/repository/domain/user/UserDao_Impl : io/getstream/chat/android/livedata/repository/domain/user/UserDao {
+	public fun <init> (Landroidx/room/RoomDatabase;)V
+	public fun insert (Lio/getstream/chat/android/livedata/repository/domain/user/UserEntity;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun insertMany (Ljava/util/List;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun select (Ljava/lang/String;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
+	public fun select (Ljava/util/List;Lkotlin/coroutines/Continuation;)Ljava/lang/Object;
 }
 
 public abstract interface class io/getstream/chat/android/livedata/usecase/CancelMessage {
