+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/domain/user/UserDao.kt
@@ -1,10 +1,9 @@
-package io.getstream.chat.android.livedata.dao
+package io.getstream.chat.android.livedata.repository.domain.user
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import io.getstream.chat.android.livedata.entity.UserEntity
 
 @Dao
 internal interface UserDao {
